+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Improved `piral debug` to respect `--public-url` like `piral build`
 - Fixed `publicPath` in pilets built with Webpack to be `./` instead of `/`
 - Fixed default imports of shared dependencies with `v2` schema (#433)
+- Fixed usage of `rules` in `extend-config` of `piral-cli-webpack(5)`
 - Fixed usage of custom public path with `piral-ng` (#434)
 - Removed `file-loader` from `piral-cli-webpack5`
 - Added env variable `PIRAL_PUBLIC_PATH` during `piral build` and `piral debug` (#434)